refactor: move models and utils to top level
will reuse these between src for generator/api and tests now
