+++ models/dao-token-contract.ts
@@ -1,6 +1,6 @@
 import { ContractBase } from "./contract-template";
-import { DEPLOYMENT_ORDER } from "../utilities/contract-deployment-order";
-import { ContractSubtype } from "../utilities/contract-types";
+import { DEPLOYMENT_ORDER } from "../../utilities/contract-deployment-order";
+import { ContractSubtype } from "../../utilities/contract-types";
 
 export class TokenContract extends ContractBase {
   constructor(name: string, subtype: ContractSubtype<"TOKEN">) {
