+++ tests/dao/extensions/aibtc-dao-charter.test.ts
@@ -1,8 +1,8 @@
 import { Cl } from "@stacks/transactions";
 import { describe, expect, it } from "vitest";
-import { ErrCodeDaoCharter } from "../../utilities/contract-error-codes";
-import { setupDaoContractRegistry } from "../../utilities/contract-registry";
-import { constructDao } from "../../utilities/dao-helpers";
+import { ErrCodeDaoCharter } from "../../../utilities/contract-error-codes";
+import { setupDaoContractRegistry } from "../../../utilities/contract-registry";
+import { constructDao } from "../../../utilities/dao-helpers";
 
 // setup accounts
 const accounts = simnet.getAccounts();
