+++ tests/dao/extensions/aibtc-treasury.test.ts
@@ -1,9 +1,8 @@
 import { Cl } from "@stacks/transactions";
 import { describe, expect, it } from "vitest";
-import { ErrCodeTreasury } from "../../utilities/contract-error-codes";
-import { setupDaoContractRegistry } from "../../utilities/contract-registry";
-import { constructDao, getDaoTokens } from "../../utilities/dao-helpers";
-import { SBTC_CONTRACT } from "../../utilities/contract-helpers";
+import { ErrCodeTreasury } from "../../../utilities/contract-error-codes";
+import { setupDaoContractRegistry } from "../../../utilities/contract-registry";
+import { getDaoTokens } from "../../../utilities/dao-helpers";
 
 // setup accounts
 const accounts = simnet.getAccounts();
