+++ tests/dao/token/aibtc-pre-faktory.test.ts
@@ -1,7 +1,7 @@
 import { Cl, ClarityType, cvToValue, UIntCV } from "@stacks/transactions";
 import { describe, expect, it } from "vitest";
-import { setupDaoContractRegistry } from "../../utilities/contract-registry";
-import { getDaoTokens } from "../../utilities/dao-helpers";
+import { setupDaoContractRegistry } from "../../../utilities/contract-registry";
+import { getDaoTokens } from "../../../utilities/dao-helpers";
 
 // setup accounts
 const accounts = simnet.getAccounts();
