+++ utilities/contract-helpers.ts
