+++ utilities/contract-traits.ts
